# internal/delivery/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/chat/trainer": {
            "get": {
                "description": "Get all chats for a trainer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chats"
                ],
                "summary": "Get Trainer Chats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of chats",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Chat"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad JWT provided",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "JWT is expired or invalid",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/chat/trainer/{user_id}": {
            "get": {
                "description": "Get messages for a chat between a user and a trainer - trainer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chats"
                ],
                "summary": "Get Chat Messages Trainer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Cursor for pagination",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of messages with pagination",
                        "schema": {
                            "$ref": "#/definitions/dto.MessagePagination"
                        }
                    },
                    "400": {
                        "description": "Bad path or JWT provided",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "JWT is expired or invalid",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/chat/user": {
            "get": {
                "description": "Get all chats for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chats"
                ],
                "summary": "Get User Chats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of chats",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Chat"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad JWT provided",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "JWT is expired or invalid",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/chat/user/{trainer_id}": {
            "get": {
                "description": "Get messages for a chat between a user and a trainer - user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chats"
                ],
                "summary": "Get Chat Messages User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Trainer ID",
                        "name": "trainer_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Cursor for pagination",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of messages with pagination",
                        "schema": {
                            "$ref": "#/definitions/dto.MessagePagination"
                        }
                    },
                    "400": {
                        "description": "Bad path or JWT provided",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "JWT is expired or invalid",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/api/service/{id}": {
            "get": {
                "description": "Get a service by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Get Service by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service details",
                        "schema": {
                            "$ref": "#/definitions/dto.BasePrice"
                        }
                    },
                    "400": {
                        "description": "Invalid service ID",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Service not found",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BasePrice": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
...
        "dto.Chat": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_message": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "time_last_message": {
                    "type": "string"
                }
            }
        },
...
        "dto.Message": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_to_user": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "service_id": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                },
                "trainer_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "dto.MessagePagination": {
            "type": "object",
            "properties": {
                "cursor": {
                    "type": "integer"
                },
                "objects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Message"
                    }
                }
            }
        },
...
        "responses.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}